Move StepsList to appropriate folder
