+++ app/src/pages/inside/common/testCaseList/stepsList/stepsList.scss
